+++ src/main/java/com/ametrinstudios/ametrin/world/block/FlammableSlabPlankBlock.java
@@ -3,18 +3,16 @@ package com.ametrinstudios.ametrin.world.block;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.level.BlockGetter;
-import net.minecraft.world.level.block.RotatedPillarBlock;
+import net.minecraft.world.level.block.SlabBlock;
 import net.minecraft.world.level.block.state.BlockState;
 
-public class FlammableRotatedPillarBlock extends RotatedPillarBlock {
-    public FlammableRotatedPillarBlock(Properties properties) {
-        super(properties);
-    }
+public class FlammableSlabPlankBlock extends SlabBlock {
+    public FlammableSlabPlankBlock(Properties properties) {super(properties);}
 
     @Override
     public boolean isFlammable(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return true;}
     @Override
     public int getFlammability(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 5;}
     @Override
-    public int getFireSpreadSpeed(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 5;}
+    public int getFireSpreadSpeed(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 20;}
 }